+++ src/main/java/com/lowdragmc/mbd2/integration/kubejs/events/MBDClientEvents.java
@@ -1,6 +1,7 @@
 package com.lowdragmc.mbd2.integration.kubejs.events;
 
 import com.lowdragmc.mbd2.MBD2;
+import com.lowdragmc.mbd2.api.recipe.event.FuelRecipeUIEvent;
 import com.lowdragmc.mbd2.api.recipe.event.RecipeTypeEvent;
 import com.lowdragmc.mbd2.common.machine.definition.config.event.*;
 import dev.latvian.mods.kubejs.event.EventHandler;
@@ -11,13 +12,13 @@ import org.jetbrains.annotations.Nullable;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
-import java.util.function.Consumer;
 import java.util.function.Function;
 
 import static com.lowdragmc.mbd2.integration.kubejs.events.MBDMachineEvents.MBD_MACHINE_EVENTS;
 
 public interface MBDClientEvents {
-    Map<Class<? extends MachineEvent>, Function<MachineEvent, EventResult>> eventHandlers = new HashMap<>();
+    Map<Class<? extends MachineEvent>, Function<MachineEvent, EventResult>> machineEventHandlers = new HashMap<>();
+    Map<Class<? extends RecipeTypeEvent>, Function<RecipeTypeEvent, EventResult>> recipeTypeEventHandlers = new HashMap<>();
 
     // Client events
     EventHandler CLIENT_TICK = registerMachineEvent("onClientTick",
@@ -32,6 +33,17 @@ public interface MBDClientEvents {
 
     @Nullable
     EventHandler CUSTOM_KEYFRAME = createCustomKeyframeEvent();
+
+    // Recipe events
+    EventHandler RECIPE_UI = registerRecipeTypeEvent("onRecipeUI",
+            RecipeTypeEvent.class,
+            MBDRecipeTypeEvents.RecipeUIEventJS.class,
+            MBDRecipeTypeEvents.RecipeUIEventJS::new);
+
+    EventHandler FUEL_RECIPE_UI = registerRecipeTypeEvent("onFuelRecipeUI",
+            FuelRecipeUIEvent.class,
+            MBDRecipeTypeEvents.FuelRecipeUIEventJS.class,
+            MBDRecipeTypeEvents.FuelRecipeUIEventJS::new);
 
     static EventHandler createCustomKeyframeEvent() {
         if (MBD2.isGeckolibLoaded()) {
@@ -51,16 +63,23 @@ public interface MBDClientEvents {
                                                                       Class<? extends MBDMachineEvents.MachineEventJS<E>> eventJSClass,
                                                                       Function<E, MBDMachineEvents.MachineEventJS<E>> eventJSFactory) {
         var handler = MBD_MACHINE_EVENTS.client(name, () -> eventJSClass).extra(Extra.ID);
-        eventHandlers.put(eventClass, event -> handler.post(eventJSFactory.apply((E) event), event.machine.getDefinition().id()));
+        machineEventHandlers.put(eventClass, event -> handler.post(eventJSFactory.apply((E) event), event.machine.getDefinition().id()));
+        return handler;
+    }
+
+    static <E extends RecipeTypeEvent> EventHandler registerRecipeTypeEvent(String name, Class<? extends RecipeTypeEvent> eventClass,
+                                                                            Class<? extends MBDRecipeTypeEvents.RecipeTypeEventJS<E>> eventJSClass,
+                                                                            Function<E, MBDRecipeTypeEvents.RecipeTypeEventJS<E>> eventJSFactory) {
+        var handler = MBDRecipeTypeEvents.MBD_RECIPE_TYPE_EVENTS.server(name, () -> eventJSClass).extra(Extra.ID);
+        recipeTypeEventHandlers.put(eventClass, event -> handler.post(eventJSFactory.apply((E) event), event.recipeType.getRegistryName()));
         return handler;
     }
     
     static EventResult postMachineEvent(MachineEvent machineEvent) {
-        return Optional.ofNullable(eventHandlers.get(machineEvent.getClass())).map(handler -> handler.apply(machineEvent)).orElse(EventResult.PASS);
+        return Optional.ofNullable(machineEventHandlers.get(machineEvent.getClass())).map(handler -> handler.apply(machineEvent)).orElse(EventResult.PASS);
     }
 
     static EventResult postRecipeTypeEvent(RecipeTypeEvent recipeTypeEvent) {
-        // TODO: Implement this method if we have client side recipe type event
-        return EventResult.PASS;
+        return Optional.ofNullable(recipeTypeEventHandlers.get(recipeTypeEvent.getClass())).map(handler -> handler.apply(recipeTypeEvent)).orElse(EventResult.PASS);
     }
 }