+++ src/main/java/com/lowdragmc/mbd2/integration/kubejs/events/MBDRecipeTypeEvents.java
@@ -1,6 +1,8 @@
 package com.lowdragmc.mbd2.integration.kubejs.events;
 
+import com.lowdragmc.mbd2.api.recipe.event.FuelRecipeUIEvent;
 import com.lowdragmc.mbd2.api.recipe.event.RecipeTypeEvent;
+import com.lowdragmc.mbd2.api.recipe.event.RecipeUIEvent;
 import com.lowdragmc.mbd2.api.recipe.event.TransferProxyRecipeEvent;
 import dev.latvian.mods.kubejs.event.EventGroup;
 import dev.latvian.mods.kubejs.event.EventJS;
@@ -20,6 +22,18 @@ public class MBDRecipeTypeEvents {
 
     public static class TransferProxyRecipeEventJS extends RecipeTypeEventJS<TransferProxyRecipeEvent> {
         public TransferProxyRecipeEventJS(TransferProxyRecipeEvent event) {
+            super(event);
+        }
+    }
+
+    public static class RecipeUIEventJS extends RecipeTypeEventJS<RecipeUIEvent> {
+        public RecipeUIEventJS(RecipeUIEvent event) {
+            super(event);
+        }
+    }
+
+    public static class FuelRecipeUIEventJS extends RecipeTypeEventJS<FuelRecipeUIEvent> {
+        public FuelRecipeUIEventJS(FuelRecipeUIEvent event) {
             super(event);
         }
     }