+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## v1.0.36
 * Fixed cannot open ui while rejoining the world
 * Fixed liquid bucket consumed while using the multiblock builder
+* Added event to modify recipe/fuel recipe ui